some more dry up
