+++ openapi/provider_test.go
@@ -267,18 +267,8 @@ definitions:
 						resourceName := fmt.Sprintf("%s_cdn_datasource_v1", providerName)
 						So(tfProvider.DataSourcesMap, ShouldContainKey, resourceName)
 
-						// TODO: dry out the assertions
-						So(tfProvider.DataSourcesMap[resourceName].Schema, ShouldContainKey, "label")
-						So(tfProvider.DataSourcesMap[resourceName].Schema["label"].Type, ShouldEqual, schema.TypeString)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["label"].Required, ShouldBeFalse)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["label"].Optional, ShouldBeTrue)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["label"].Computed, ShouldBeTrue)
-
-						So(tfProvider.DataSourcesMap[resourceName].Schema, ShouldContainKey, "owners")
-						So(tfProvider.DataSourcesMap[resourceName].Schema["owners"].Type, ShouldEqual, schema.TypeList)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["owners"].Required, ShouldBeFalse)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["owners"].Optional, ShouldBeTrue)
-						So(tfProvider.DataSourcesMap[resourceName].Schema["owners"].Computed, ShouldBeTrue)
+						assertDataSourceSchemaProperty(t, tfProvider.DataSourcesMap[resourceName].Schema["label"], schema.TypeString)
+						assertDataSourceSchemaProperty(t, tfProvider.DataSourcesMap[resourceName].Schema["owners"], schema.TypeList)
 
 						So(tfProvider.DataSourcesMap[resourceName].Schema, ShouldContainKey, "filter")
 						So(tfProvider.DataSourcesMap[resourceName].Schema["filter"].Type, ShouldEqual, schema.TypeSet)
